+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to the desktop application will be documented in this file.
 
 ### Bug Fixes
 
-- Keep tray window reachable from dock (1e227c3)
+- Make Windows shortcuts usable across platforms (1b0e6fc)
+- Keep tray window reachable from dock (b53ced4)
 
 ## v0.6.3 (2026-05-25)